chore: try catch ton wallets fetch
